+++ src/retrocookie/utils.py
@@ -1,11 +1,8 @@
 """Utilities."""
 import contextlib
 import os
-import tempfile
 from pathlib import Path
 from typing import Iterator
-
-from . import git
 
 
 def removeprefix(string: str, prefix: str) -> str:
@@ -28,11 +25,3 @@ def chdir(path: Path) -> Iterator[None]:
         yield
     finally:
         os.chdir(cwd)
-
-
-@contextlib.contextmanager
-def temporary_repository() -> Iterator[git.Repository]:
-    """Create repository in temporary directory."""
-    with tempfile.TemporaryDirectory() as tmpdir:
-        path = Path(tmpdir)
-        yield git.Repository.init(path)
